services/atlas/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
psycopg2-binary
pyjwt
python-dotenv
gunicorn
alembic
httpx
openai
bcrypt
python-multipart
email-validator
selenium
webdriver-manager
pillow